Migrate blog to appdir (WIP)
