+++ app/blog/page.tsx
@@ -1,6 +1,4 @@
 import React from 'react'
-import {NextPage} from "next";
-import Layout from "../../components/Layout";
 import Title from "../../components/Title";
 import {BlogPost, getSortedPostsData} from "../../lib/blog/load";
 
@@ -10,32 +8,21 @@ import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {faStar} from "@fortawesome/free-solid-svg-icons";
 import ArticleCard, {ArticleGrid} from "../../components/blog/ArticleCard";
 
-export const getStaticProps = async () => {
+
+export default async function Index() {
   const articles = await getSortedPostsData()
   const latestLongArticleIdx = articles.findIndex((e: BlogPost) => e.tags.includes("長編記事"))
   const latestLongArticle = articles[latestLongArticleIdx];
   articles.splice(latestLongArticleIdx, 1);
-  return {
-    props: {
-      latestLongArticle,
-      otherArticles: articles
-    }
-  }
-}
+  const otherArticles = articles
 
-type Props = {
-  latestLongArticle: BlogPost,
-  otherArticles: BlogPost[]
-};
-
-const Blog: NextPage<Props> = ({latestLongArticle, otherArticles}) => {
   const description =
     'つまみさんのブログです。' +
     '主にお散歩やソフトウェアの記事を書いています。'
 
   return (
     <>
-      <Layout>
+      <div id="main_wrapper">
         <Title title={'つまみログ'} description={description}/>
 
         <div className={styles.hrule_block}>
@@ -53,9 +40,7 @@ const Blog: NextPage<Props> = ({latestLongArticle, otherArticles}) => {
         <ArticleGrid>
           {otherArticles.map(entry => <ArticleCard entry={entry} key={entry.slug}/>)}
         </ArticleGrid>
-      </Layout>
+      </div>
     </>
   )
 }
-
-export default Blog
