+++ next.config.js
@@ -71,28 +71,28 @@ module.exports = withMdx(config)
  * Source: https://miyaoka.dev/posts/2020-12-31-hmr-on-markdown
  */
 
-const fs = require('fs');
-const chokidar = require('chokidar');
-
-let isUpdating  = false
-const postsDir  = './posts'
-const scriptDir = './lib/blog/fileWatch.ts'
-
-const handler = () => {
-  if (isUpdating) return
-  isUpdating = true
-  const content = fs.readFileSync(scriptDir, 'utf-8')
-  const codeToInsert = `${content}\n console.log()`
-
-  fs.writeFileSync(scriptDir, codeToInsert)
-
-  setTimeout(() => {
-    fs.writeFileSync(scriptDir, content)
-    isUpdating = false
-  }, 1000)
-}
-
-chokidar
-  .watch(postsDir, { ignoreInitial: true })
-  .on('add', handler)
-  .on('change', handler);
+// const fs = require('fs');
+// const chokidar = require('chokidar');
+//
+// let isUpdating  = false
+// const postsDir  = './posts'
+// const scriptDir = './lib/blog/fileWatch.ts'
+//
+// const handler = () => {
+//   if (isUpdating) return
+//   isUpdating = true
+//   const content = fs.readFileSync(scriptDir, 'utf-8')
+//   const codeToInsert = `${content}\n console.log()`
+//
+//   fs.writeFileSync(scriptDir, codeToInsert)
+//
+//   setTimeout(() => {
+//     fs.writeFileSync(scriptDir, content)
+//     isUpdating = false
+//   }, 1000)
+// }
+//
+// chokidar
+//   .watch(postsDir, { ignoreInitial: true })
+//   .on('add', handler)
+//   .on('change', handler);
